JsonData type and friends.
